Finished Scrolling/Masking Quarters
Finished making the quarter system work well.  It scrolls nicely to view
everything, and is masked so that it looks nice too.  I'd say this is a
good version 1.2.  The last thing I want to implement is an expander and
an overall showing for each quarter.  Then I need to organize it by year
and by date.